requirements.txt
# Core dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
supabase==2.5.0
pgvector==0.4.2

# Scraping
scrapling==0.4.1
DrissionPage==4.1.1.2
playwright==1.58.0
patchright==1.58.0

# Scheduling
APScheduler==3.11.2

# Telegram
python-telegram-bot==22.6

# LLM/Embeddings
google-generativeai==0.8.6

# Video/Audio
elevenlabs==1.54.0

# Payments/Email
resend==2.0.0
stripe==9.4.0

# ML/Analytics
lightgbm==4.3.0
pandas==3.0.1
numpy==2.4.2
scikit-learn==1.8.0
scipy==1.17.1

# Auth/JWT
python-jose[cryptography]

# Utilities (used by API / scraping)
httpx
pydantic
python-multipart
jinja2
aiohttp
beautifulsoup4
lxml
curl_cffi
msgspec
browserforge
ffmpeg-python
